api/requirements.txt
# Core web framework
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1

# Configuration and database
dj-database-url==2.2.0
psycopg[binary]==3.2.3

# Production serving
gunicorn==22.0.0
cryptography==44.0.2

# Reporting and document generation
python-dateutil==2.8.2
reportlab==4.0.9
Pillow==12.1.1
